search_sample/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
SQLAlchemy==2.0.28
psycopg2-binary==2.9.9
azure-storage-blob==12.19.1
azure-servicebus==7.12.1
azure-eventgrid==4.11.0
azure-identity==1.16.0
azure-core==1.30.0
pymongo==4.6.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
PyMuPDF==1.23.5 
openai==1.10.0
python-dotenv==1.0.1
loguru==0.7.2
PyYAML==6.0.1
pytesseract==0.3.10
Pillow==9.5.0
python-docx==0.8.11
